move email capture to its own folder, move helpers to their own file, add live calls to formbricks and error handling, add saving to session storage
